Clear google-chrome's LD_LIBRARY_PATH before spawning the player
Google Chrome ships with bundled libvulkan.so.1 and links to it by
setting LD_LIBRARY_PATH in the startup script.

When acestream-launcher is invoked from chrome through the
acestream:// protocol, the environment variables are passed in
and eventually passed to the player.

In fedora + rpmfusion, when mpv is spawned it fails to link because
it attempts to link against google's build of libvulkan instead of
the system provided library.